style: apply rustfmt + clippy fixes under the pinned toolchain
Mechanical pass to make `cargo fmt --all --check` and
`cargo clippy --workspace -- -D warnings` green under the newly pinned
1.94.0 toolchain. No runtime behavior changes.

- rustfmt reflow across the workspace (bulk of the diff).
- Fix genuine clippy lints: unnecessary_unwrap, question_mark/let-else,
  needless_option_as_deref, field_reassign_with_default, while_let_loop,
  needless_range_loop, unused_parens, doc_lazy_continuation.
- Allow architectural/noise lints at the narrowest scope with a comment:
  too_many_arguments, manual_c_str_literals (dx.rs module allow),
  missing_transmute_annotations, large_enum_variant, new_ret_no_self.
- Drop unused core-text/core-foundation imports in ui_shaper.rs that
  broke the macOS build after the upstream dev merge.

+++ crates/loom-term/src/sixel.rs
@@ -208,7 +208,7 @@ fn sixel_band_height(sixel: u8) -> u32 {
     if sixel == 0 {
         0
     } else {
-        (u8::BITS - sixel.leading_zeros()) as u32
+        u8::BITS - sixel.leading_zeros()
     }
 }
